store generators instead of indices
Quoting from Exercise 6 in the notebook:

In schreier we store indices (in S) of generators in the dictionary, so
we need to keep the set around. It is not necessary though: since julia
passes arguments by reference and Permutations are heap-allocated there
can be more than one reference to each Permutation at a given time. We
can exploit it this feature to store directly references to generators
in the dictionary. How? Just say Sch[γ] = s instead of Sch[γ] = idx.